pyproject.toml
[project]
name = "loja-db"
version = "0.1.0"
description = ""
authors = [
    {name = "m4rco0",email = "m4rco.unknow1011@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, < 4.0"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.1,<0.16.0)"
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev loja_db/app.py'
pre_test = 'task lint'
test = 'pytest -s -x  --cov=loja_db -vv'
post_test = 'coverage html; firefox htmlcov/index.html &'
